Retain and reuse matching vnode from vfs hash table
If a vnode with same hash, mount & qid is found in the hash table, while
inserting another vnode, it was incorrectly considered an error (without
'error' being set) and vnode pointer was set to NULL. That is not really
an error. It is expected to retain and reuse the existing vnode from the
vfs hash table, if it has same hash, mount and qid.

Change-Id: I112a55983f61873e4e6fd323953fc29720178476